+++ src/components/atomes/index.ts
@@ -1,4 +1,3 @@
-export { default as Card } from './card';
 export { default as Button } from './button';
 export { default as Flex } from './flex';
 export { default as DivideLine } from './divide-line';
