+++ src/components/molecules/card/Info.tsx
