+++ src/components/molecules/index.ts
@@ -1,3 +1,4 @@
 export { default as Header } from './header';
 export { default as Modal } from './modal';
+export { default as Card } from './card';
